SFOS-1617
Review and clean up some of the display code

+++ trunk/core/smartfrog/src/org/smartfrog/services/management/OptionSet.java
@@ -117,7 +117,7 @@ public class OptionSet {
                             }
 
                             isRemoteSubprocess = true;
-                            port = new Integer(args[++i]).intValue();
+                            port = new Integer(args[++i]);
 
                             break;
